Solve task 5
I placed the Season method in the Program class. It could have been in
the Student class, but ideally it would be a property of the DateTime
class. It didn't seem right to place it in the Student class and I did
not want to subclass DateTime, so I simply made it a static method in
the Program class.

The Season method takes a Student as its argument, but a better, more
general solution would have been to make it take a month number instead.
It returns a hard-coded string for each season, but it would probably be
a better solution to create an enum with values for each season and
return that instead.