Solve task 6
Also, handle out-of-range parameters in Season method from task 5.

PrintBirthdaysPerSeason implements the solution for task 6. I made it a
static method in the Program class. It could have been made an instance
method of the ClassRoom class, but this seemed messy to me.

It implements very specific functionality that is not essential to the
ClassRoom type, so it makes sense to make it a simple method in the
Program class. The ClassRoom type could be moved to its own assembly and
be reused in other programs, but this specific functionality is less
likely to be reused.

However, it could potentially be reused for other groups of students
than those in a ClassRoom, so it makes sense to write to accept the
minimal necessary information: an IEnumerable<Student>. This design has
lower coupling and superior reusability compared to a design where the
printing method is in the ClassRoom class.